+++ python/cudf/tests/test_dataframe.py
@@ -381,7 +381,9 @@ def test_dataframe_slicing():
     del subrange
 
 
-def test_dataframe_loc():
+@pytest.mark.parametrize('step', [1, 2, 5])
+@pytest.mark.parametrize('scalar', [0, 20, 100])
+def test_dataframe_loc(scalar, step):
     df = DataFrame()
     size = 123
     df['a'] = ha = np.random.randint(low=0, high=100, size=size)\
@@ -391,17 +393,34 @@ def test_dataframe_loc():
         .astype(np.int64)
     df['d'] = hd = np.random.random(size).astype(np.float64)
 
+    pdf = pd.DataFrame()
+    pdf['a'] = ha
+    pdf['b'] = hb
+    pdf['c'] = hc
+    pdf['d'] = hd
+
+    # Scalar label
+    np.testing.assert_equal(df.loc[scalar].to_array(), pdf.loc[scalar])
+
     # Full slice
     full = df.loc[:, ['c']]
     assert tuple(full.columns) == ('c',)
     np.testing.assert_equal(full['c'].to_array(), hc)
 
-    begin = 117
+    begin = 110
     end = 122
-    fewer = df.loc[begin:end, ['c', 'd', 'a']]
-    assert len(fewer) == end - begin + 1
+
+    fewer = df.loc[begin:end:step, ['c', 'd', 'a']]
+    assert len(fewer) == (end - begin)//step + 1
     assert tuple(fewer.columns) == ('c', 'd', 'a')
-    np.testing.assert_equal(fewer['a'].to_array(), ha[begin:end + 1])
+    np.testing.assert_equal(fewer['a'].to_array(), ha[begin:end + 1:step])
+    np.testing.assert_equal(fewer['c'].to_array(), hc[begin:end + 1:step])
+    np.testing.assert_equal(fewer['d'].to_array(), hd[begin:end + 1:step])
+    del fewer
+
+    fewer = df.loc[begin:end, ['c', 'd']]
+    assert len(fewer) == end - begin + 1
+    assert tuple(fewer.columns) == ('c', 'd')
     np.testing.assert_equal(fewer['c'].to_array(), hc[begin:end + 1])
     np.testing.assert_equal(fewer['d'].to_array(), hd[begin:end + 1])
     del fewer
@@ -417,6 +436,24 @@ def test_dataframe_loc():
     np.testing.assert_equal(fewer['a'].to_array(), ha[begin:end + 1])
     np.testing.assert_equal(fewer['c'].to_array(), hc[begin:end + 1])
     np.testing.assert_equal(fewer['d'].to_array(), hd[begin:end + 1])
+
+
+@pytest.mark.xfail(
+    raises=IndexError,
+    reason="label scalar is out of bound"
+)
+def test_dataframe_loc_outbound():
+    df = DataFrame()
+    size = 10
+    df['a'] = ha = np.random.randint(low=0, high=100, size=size) \
+        .astype(np.int32)
+    df['b'] = hb = np.random.random(size).astype(np.float32)  # noqa: F841
+
+    pdf = pd.DataFrame()
+    pdf['a'] = ha
+    pdf['b'] = hb
+
+    np.testing.assert_equal(df.loc[11].to_array(), pdf.loc[11])
 
 
 @pytest.mark.parametrize('nelem', [2, 5, 20, 100])